refactor: bottomSheetModalProvider 최상위 컴포넌트에 선언
