+++ src/components/SelectRegion/SidoTable.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Pressable, StyleSheet, Text, View, FlatList } from 'react-native';
-import { fetchSigunguInfo } from '../../apis/fetchSigunguInfo';
+import { fetchSigunguInfo } from '../../apis/fetchRegionAPI/fetchSigunguInfo';
 import COLOR from '../../constants/colors';
 import { SidoType, SigunguAndEupmyeondongType, SigunguFeatureType } from './types';
